Fix queries, modified functions, and various bug fixes

+++ app/models/bookmarks.php
@@ -2,7 +2,7 @@
 
 class Bookmarks extends AppModel
 {
-    public static function setBookmarks($user_id, $thread_id)
+    public static function add($user_id, $thread_id)
     {
         $db = DB::conn();
 
@@ -14,20 +14,18 @@ class Bookmarks extends AppModel
         $db->insert('bookmark', $params);
     }
 
-    public static function unsetBookmarks($user_id, $thread_id)
+    public static function remove($user_id, $thread_id)
     {
         $db = DB::conn();
 
-        $db->query('DELETE FROM bookmark WHERE user_id = ? AND thread_id = ?',
-        array($user_id, $thread_id));
+        $db->query('DELETE FROM bookmark WHERE user_id = ? AND thread_id = ?', array($user_id, $thread_id));
     }
 
     public static function isBookmark($user_id, $thread_id)
     {
         $db = DB::conn();
 
-        $row = $db->row('SELECT * FROM bookmark WHERE user_id = ? AND thread_id = ?',
-        array($user_id, $thread_id));
+        $row = $db->row('SELECT * FROM bookmark WHERE user_id = ? AND thread_id = ?', array($user_id, $thread_id));
 
         return (bool) $row;
     }
@@ -37,8 +35,7 @@ class Bookmarks extends AppModel
         $threads = array();
         $db = DB::conn();
 
-        $rows = $db->rows('SELECT thread_id FROM bookmark WHERE user_id = ?',
-        array($user_id));
+        $rows = $db->rows('SELECT thread_id FROM bookmark WHERE user_id = ?', array($user_id));
 
         foreach ($rows as $row) {
             $threads[] = Thread::get($row['thread_id']);  //pass the thread_id value
